First prototype of the SLM handling events from infrastructure adapter.
The SLM assumes the infrastructure adapter response contains a NSR and one VNFR for each VNF. The SLM includes those VNFRs in the NSR, generates an id if it doesn't have one, and send it to the Network Service repository to save it in the database.

Still code contains some "TODOs".