SLM asks for available resources before sending NSD and VNFDs to Infrastructure Adaptor.
The callback method for the get resources notification required access to the NSD and VNFD in order to send it to IA to deploy a service, but these information was not avaiable in this thread. Therefore, a closure has been included. It would make possible that the callback method can access NSD and VNFDs coming from GK.